Re-organized search components.
