+++ src/components/search/peptideSearchBox/helpers/form/MultiQueryAlignmentOptionsForm.tsx
