+++ src/components/search/peptideSearchBox/helpers/form/SingleQueryAlignmentOptionsForm.tsx
