+++ src/components/search/peptideSearchBox/helpers/help/RegexHelpMessage.tsx
