+++ src/components/search/peptideSearchBox/helpers/input/MetadataFiltersFormItem.tsx
