+++ src/components/search/peptideSearchBox/helpers/input/TextSearchInput.tsx
