+++ src/components/search/peptideSearchBox/tabs/SingleQueryPeptideSearchBox.tsx
@@ -5,8 +5,8 @@ import Button from 'react-bootstrap/Button';
 import Spinner from 'react-bootstrap/Spinner';
 import { useRouter } from 'next/navigation';
 import { ErrorMessage } from '@components/common/errorMessage';
-import { SingleQueryAlignmentOptionsForm } from '../helpers/SingleQueryAlignmentOptionsForm';
-import { TextSearchInput } from '../helpers/TextSearchInput';
+import { SingleQueryAlignmentOptionsForm } from '../helpers/form/SingleQueryAlignmentOptionsForm';
+import { TextSearchInput } from '../helpers/input/TextSearchInput';
 import { postSingleQuerySearchAction } from '@actions/search/alignment/single-query';
 import MagnifyingGlassIcon from '@assets/svg/icons/magnifying-glass-solid.svg';
 import { RouteDefs } from '@lib/constants/routes';
