+++ src/components/search/peptideSearchBox/tabs/TextQueryPeptideSearchBox.tsx
@@ -4,12 +4,12 @@ import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
 import Spinner from 'react-bootstrap/Spinner';
 import { useRouter } from 'next/navigation';
-import { TextSearchInput } from '../helpers/TextSearchInput';
-import { RegexHelpMessage } from '../helpers/RegexHelpMessage';
-import { FiltersHelpMessage } from '../helpers/FiltersHelpMessage';
-import { SequenceLengthFilterForm } from '../helpers/SequenceLengthFilterForm';
-import { MetadataFiltersForm } from '../helpers/MetadataFiltersForm';
-import { AttributesFiltersForm } from '../helpers/AttributesFiltersForm';
+import { TextSearchInput } from '../helpers/input/TextSearchInput';
+import { RegexHelpMessage } from '../helpers/help/RegexHelpMessage';
+import { FiltersHelpMessage } from '../helpers/help/FiltersHelpMessage';
+import { SequenceLengthFilterForm } from '../helpers/form/SequenceLengthFilterForm';
+import { MetadataFiltersForm } from '../helpers/form/MetadataFiltersForm';
+import { AttributesFiltersForm } from '../helpers/form/AttributesFiltersForm';
 import MagnifyingGlassIcon from '@assets/svg/icons/magnifying-glass-solid.svg';
 import { RouteDefs } from '@lib/constants/routes';
 import { TextQueryMetadataFilter, TextQueryAttributeFilter, SequenceLengthFilter, convertMetadataFilterToParam, convertAttributeFilterToParam, convertSequenceLengthFilterToParam } from '@lib/services/api/models/search';
